tidying

+++ src/src/tls-openssl.c
@@ -2576,8 +2576,8 @@ SSL *ssl = is_server ? server_ssl : client_ssl;
 static uschar * corked = NULL;
 static int c_size = 0, c_len = 0;
 
-DEBUG(D_tls) debug_printf("%s(%p, %d%s)\n", __FUNCTION__,
-  buff, len, more ? ", more" : "");
+DEBUG(D_tls) debug_printf("%s(%p, %lu%s)\n", __FUNCTION__,
+  buff, (ulong)len, more ? ", more" : "");
 
 /* Lacking a CORK or MSG_MORE facility (such as GnuTLS has) we copy data when
 "more" is notified.  This hack is only ok if small amounts are involved AND only
